requirements.txt
crewai==0.1.30
langchain-openai==0.0.5
langchain-community==0.0.20
python-dotenv==1.0.1
flask==3.0.2
duckduckgo-search==4.4.3
wikipedia-api==0.6.0
beautifulsoup4==4.12.3
requests==2.31.0
openai==1.12.0
langchain==0.1.9
langchain-core==0.1.27 